fix(workloads): decouple runtime health from convergence
